+++ src/main/java/com/bae/persistence/repository/RoleDatabaseRepository.java
@@ -47,6 +47,9 @@ public class RoleDatabaseRepository implements RoleRepository {
 	@Override
 	@Transactional(REQUIRED)
 	public String deleteRole(int id) {
+		if (!checkRoleExists(id)) {
+			return ROLE_NOT_FOUND;
+		}
 		Role role = entityManager.find(Role.class, id);
 		entityManager.remove(role);
 		return DELETE_ROLE_SUCCESS;
@@ -55,13 +58,16 @@ public class RoleDatabaseRepository implements RoleRepository {
 	@Override
 	@Transactional(REQUIRED)
 	public String updateRole(int id, String role) {
-		Role updatedRole = util.getObjectForJSON(role, Role.class);
-
 		if (!checkRoleExists(id)) {
 			return ROLE_NOT_FOUND;
 		}
 
-		entityManager.merge(updatedRole);
+		Role updatedRole = util.getObjectForJSON(role, Role.class);
+		Role oldRole = entityManager.find(Role.class, id);
+
+		oldRole.setName(updatedRole.getName());
+
+		entityManager.merge(oldRole);
 		return UPDATE_ROLE_SUCCESS;
 	}
 
@@ -75,12 +81,7 @@ public class RoleDatabaseRepository implements RoleRepository {
 	}
 
 	private boolean checkRoleExists(int id) {
-		// Execute a query rather than using entityManager.find/.contains to improve
-		// performance by not having to retrieve records from the database.
-
-		return (long) entityManager.createQuery(String.format("SELECT COUNT(r) FROM Role r WHERE r.role_id = '%s'", id))
-				.getSingleResult() == 1;
-
+		return entityManager.find(Role.class, id) != null;
 	}
 
 }
